refactor: normalize eslint migration file names
