+++ src/modules/User/features/Auth/api/index.ts
@@ -1,2 +1,2 @@
-export { default as LoginAPI } from './LoginAPI';
-export { default as RegistrationAPI } from './RegistrationAPI';
+export { default as LoginAPI } from './login-api';
+export { default as RegistrationAPI } from './registration-api';
