+++ src/modules/User/features/Auth/hooks/use-registration-form.ts
@@ -10,7 +10,7 @@ import {
   selectRegistrationError,
   selectRegistrationLoading,
   selectRegistrationUser,
-} from '@/modules/User/store/registrationSelectors';
+} from '@/modules/User/store/registration-selectors';
 
 
 type UseRegistrationFormResult = {
