+++ src/modules/User/store/index.ts
@@ -1,2 +1,2 @@
-export * from './loginSlice';
-export * from './registrationSlice';
+export * from './login-slice';
+export * from './registration-slice';
