+++ tests/integration/modules/User/store/login-slice.integration.test.ts
@@ -5,9 +5,9 @@ import '../../../setup';
 import API_ENDPOINTS from '@/config/apiConfig';
 import container from '@/config/DependencyInjectionConfig';
 import TOKENS from '@/config/tokens';
-import type LoginAPI from '@/modules/User/features/Auth/api/LoginAPI';
-import type RegistrationAPI from '@/modules/User/features/Auth/api/RegistrationAPI';
-import { loginReducer, loginUser, logout } from '@/modules/User/store/loginSlice';
+import type LoginAPI from '@/modules/User/features/Auth/api/login-api';
+import type RegistrationAPI from '@/modules/User/features/Auth/api/registration-api';
+import { loginReducer, loginUser, logout } from '@/modules/User/store/login-slice';
 import type { ThunkExtra } from '@/modules/User/store/types';
 
 import server from '../../../mocks/server';
