+++ tests/integration/modules/User/store/registration-slice-aborted.integration.test.ts
@@ -5,13 +5,13 @@ import '../../../setup';
 import API_ENDPOINTS from '@/config/apiConfig';
 import container from '@/config/DependencyInjectionConfig';
 import TOKENS from '@/config/tokens';
-import type LoginAPI from '@/modules/User/features/Auth/api/LoginAPI';
-import type RegistrationAPI from '@/modules/User/features/Auth/api/RegistrationAPI';
+import type LoginAPI from '@/modules/User/features/Auth/api/login-api';
+import type RegistrationAPI from '@/modules/User/features/Auth/api/registration-api';
 import {
   registrationReducer,
   registerUser,
   type RegistrationState,
-} from '@/modules/User/store/registrationSlice';
+} from '@/modules/User/store/registration-slice';
 import type { ThunkExtra } from '@/modules/User/store/types';
 
 import server from '../../../mocks/server';
