+++ tests/integration/utils/test-helpers.tsx
@@ -8,8 +8,8 @@ import { BrowserRouter } from 'react-router-dom';
 import container from '@/config/DependencyInjectionConfig';
 import TOKENS from '@/config/tokens';
 import i18n from '@/i18n';
-import type LoginAPI from '@/modules/User/features/Auth/api/LoginAPI';
-import type RegistrationAPI from '@/modules/User/features/Auth/api/RegistrationAPI';
+import type LoginAPI from '@/modules/User/features/Auth/api/login-api';
+import type RegistrationAPI from '@/modules/User/features/Auth/api/registration-api';
 import { loginReducer, registrationReducer } from '@/modules/User/store';
 import type { ThunkExtra } from '@/modules/User/store/types';
 import type { RootState } from '@/stores';
