+++ tests/unit/scripts/localization-generator.test.js
